InternsGuide_website/backend/requirements.txt
flask==2.3.2
flask-sqlalchemy==3.0.3
python-dotenv==0.19.0
pyjwt==2.3.0
werkzeug==2.3.3
flask-cors==3.0.10
flask-migrate==3.1.0
sqlalchemy==2.0.15